+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * New: Add integration with admin bar
     * New: Add admin bar entry for revealing template
     * New: Add setting to control if admin bar entry should appear (defaulted to true)
+    * New: Add `can_show_in_admin_bar()` for determining if the admin bar item should be shown
     * New: Add new screenshot
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it
 * Fix: Correct typo on plugin settings page